Share one expander, and discriminate windows before walking them
OpGraphBlockExpander was a near-verbatim fork of OpGraphDeviceOpExpander,
identical down to both multi-line rationale comments, so a fix to the
mousedown behaviour had to land twice. One presentational expander takes
the action wording, the count and a toggle; the two callers supply what
actually differed.

The window scan ran two O(length) isRangeConsumed walks before reaching
the O(1) test for whether the window can repeat at all. Hoisting it cuts
detection on a 10k-op no-repeat graph — the common case — from 40.4ms to
4.0ms median over five runs. Control flow is unchanged: every rejected
start still advances by one.
