+++ src/components/operation-graph/OpGraphDeviceOpExpander.tsx
@@ -2,7 +2,8 @@
 //
 // SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
 
-import { type MouseEvent as ReactMouseEvent, memo, useContext } from 'react';
+import { memo, useContext } from 'react';
+import OpGraphNodeExpander from './OpGraphNodeExpander';
 import { OpGraphExpansionContext } from './opGraphExpansionContext';
 
 interface OpGraphDeviceOpExpanderProps {
@@ -14,44 +15,16 @@ interface OpGraphDeviceOpExpanderProps {
 const OpGraphDeviceOpExpander = memo(({ operationId, count, isExpanded }: OpGraphDeviceOpExpanderProps) => {
     const toggleExpansion = useContext(OpGraphExpansionContext);
 
-    // React Flow reads selection off `mousedown`, so without this the expander
-    // would also select the operation and start the centring tween that goes with
-    // it — the graph would jump before it expanded. Structural navigation and
-    // selection stay separate gestures.
-    const handleMouseDown = (event: ReactMouseEvent) => event.stopPropagation();
-
-    const handleClick = (event: ReactMouseEvent) => {
-        event.stopPropagation();
-        toggleExpansion(operationId);
-    };
-
-    // Also the accessible name: `title` only supplies one when the element has no
-    // content, and the count is content — so without this the button announced as
-    // the bare number.
-    const action = isExpanded ? 'Hide device operations' : `Show ${count} device operations`;
-
     return (
-        <button
-            type='button'
-            // `nodrag`/`nopan` keep the button from doubling as a drag or pan surface.
-            className='op-graph-node-expander nodrag nopan'
-            title={action}
-            aria-label={action}
-            aria-expanded={isExpanded}
-            onMouseDown={handleMouseDown}
-            onClick={handleClick}
-        >
-            {/* Down opens, up closes — a chevron that only reported the current
-                state read as "expand further". Same glyph, rotated: `▴` sits
-                high in the em-box and the count looks dropped beside it. */}
-            <span
-                aria-hidden='true'
-                className='op-graph-node-expander-icon'
-            >
-                ▾
-            </span>
-            {count}
-        </button>
+        <OpGraphNodeExpander
+            // Also the accessible name: `title` only supplies one when the element
+            // has no content, and the count is content — so without this the button
+            // announced as the bare number.
+            action={isExpanded ? 'Hide device operations' : `Show ${count} device operations`}
+            count={count}
+            isExpanded={isExpanded}
+            onToggle={() => toggleExpansion(operationId)}
+        />
     );
 });
